Some macro cleanup.
